+++ .eslintignore
@@ -7,3 +7,5 @@ LICENSE
 README.md
 CONTRIBUTING.md
 coverage
+e2e-tests
+playwright-report
