--- .features-gen/e2e-tests/features/lmrValidation.feature.spec.js
@@ -1,67 +1,0 @@
-// Generated from: e2e-tests\features\lmrValidation.feature
-import { test } from "../../../e2e-tests/fixture/fixtures.ts";
-
-test.describe('LMR - Landlords make a report', () => {
-
-  test('LMR Test-1 Validate the above pages', { tag: ['@LmrRegressionTest'] }, async ({ Given, When, Then, And, page, pages }) => { 
-    await Given('I visit the Landlord\'s make a report Page', null, { pages }); 
-    await When('I choose to navigate to "When did the person move into your property?" page for LMR', null, { pages }); 
-    await And('I select continue', null, { page, pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Tell us when the person moved into your property" error summary', null, { pages }); 
-    await When('I enter "Tomorrow\'s date" into "When did the person move into your property?" field and page for LMR', null, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "This date must not be in the future" error summary', null, { pages }); 
-    await When('I enter "30/11/2014" into "When did the person move into your property?" field and page for LMR', null, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "This service only applies to tenancy agreements that started after 30 November 2014" error summary', null, { pages }); 
-    await When('I enter "Today\'s date" into "When did the person move into your property?" field and page for LMR', null, { pages }); 
-    await And('I select continue', null, { page, pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Enter tenant\'s full name¬Please enter the tenant\'s date of birth" error summary', null, { pages }); 
-    await When('I complete Existing tenant\'s information fields with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Full name"},{"value":"Dan HOF"}]},{"cells":[{"value":"Date of birth"},{"value":"less than 18 years ago"}]},{"cells":[{"value":"Nationality"},{"value":"Aland Islands"}]}]}}, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Date of birth entered must reflect 18 years or older" error summary', null, { pages }); 
-    await When('I complete Existing tenant\'s information fields with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Full name"},{"value":"Dan HOF"}]},{"cells":[{"value":"Date of birth"},{"value":"1*/02/1990"}]},{"cells":[{"value":"Nationality"},{"value":"Aland Islands"}]}]}}, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Please enter a valid date" error summary', null, { pages }); 
-    await When('I complete Existing tenant\'s information fields with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Full name"},{"value":"Dan HOF"}]},{"cells":[{"value":"Date of birth"},{"value":"19 years ago"}]},{"cells":[{"value":"Nationality"},{"value":""}]}]}}, { pages }); 
-    await And('I select continue', null, { page, pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Enter the property address line 1¬Enter the town/city¬Enter the postcode of the rental property" error summary', null, { pages }); 
-    await When('I complete Rental property address with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Address line 1"},{"value":"12"}]},{"cells":[{"value":"Address line 2"},{"value":"Kings"}]},{"cells":[{"value":"Town or City"},{"value":"Leeds"}]},{"cells":[{"value":"Country"},{"value":"UK"}]},{"cells":[{"value":"Postcode"},{"value":"L!2 1PP"}]}]}}, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Your postcode is not in England. If your property is not in England, you do not have to do a right to rent check" error summary', null, { pages }); 
-    await When('I complete Rental property address with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Address line 1"},{"value":"12"}]},{"cells":[{"value":"Address line 2"},{"value":"Kings"}]},{"cells":[{"value":"Town or City"},{"value":"Leeds"}]},{"cells":[{"value":"Country"},{"value":"UK"}]},{"cells":[{"value":"Postcode"},{"value":"M12 1PP"}]}]}}, { pages }); 
-    await And('I select continue', null, { page, pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Enter the landlord\'s full name¬Enter the landlord\'s email" error summary', null, { pages }); 
-    await When('I complete Landlord\'s or agent\'s information with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Landlord's or agent's full name"},{"value":"Kings"}]},{"cells":[{"value":"Company name"},{"value":""}]},{"cells":[{"value":"Email address"},{"value":"TesterTlf.com@"}]},{"cells":[{"value":"Phone number"},{"value":"-48081570192"}]}]}}, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Enter a valid email¬Enter a valid phone number" error summary', null, { pages }); 
-    await When('I complete Landlord\'s or agent\'s information with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Landlord's or agent's full name"},{"value":"Kings"}]},{"cells":[{"value":"Company name"},{"value":"HOF LTD"}]},{"cells":[{"value":"Email address"},{"value":"T@t.c"}]},{"cells":[{"value":"Phone number"},{"value":"016148081588"}]}]}}, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Enter a valid email¬Enter a valid phone number" error summary', null, { pages }); 
-    await When('I complete Landlord\'s or agent\'s information with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Landlord's or agent's full name"},{"value":"Kings"}]},{"cells":[{"value":"Company name"},{"value":"HOF LTD"}]},{"cells":[{"value":"Email address"},{"value":"TesterTlf.com"}]},{"cells":[{"value":"Phone number"},{"value":"01614808158"}]}]}}, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Enter a valid email" error summary', null, { pages }); 
-    await When('I complete Landlord\'s or agent\'s information with the below details:', {"dataTable":{"rows":[{"cells":[{"value":"Landlord's or agent's full name"},{"value":"Kings"}]},{"cells":[{"value":"Company name"},{"value":"HOF LTD"}]},{"cells":[{"value":"Email address"},{"value":"Tester@Tlf.com"}]},{"cells":[{"value":"Phone number"},{"value":"01614808158"}]}]}}, { pages }); 
-    await And('I select continue', null, { page, pages }); 
-    await And('I select submit report button', null, { pages }); 
-    await Then('I should see "There is a problem" error message displayed', null, { pages }); 
-    await And('I should see "Confirm you have read the Data Protection statement" error summary', null, { pages }); 
-  });
-
-});
-
-// == technical section ==
-
-test.use({
-  $test: [({}, use) => use(test), { scope: 'test', box: true }],
-  $uri: [({}, use) => use('e2e-tests\\features\\lmrValidation.feature'), { scope: 'test', box: true }],
-  $bddFileData: [({}, use) => use(bddFileData), { scope: "test", box: true }],
-});
-
-const bddFileData = [ // bdd-data-start
-  {"pwTestLine":6,"pickleLine":10,"tags":["@LmrRegressionTest"],"steps":[{"pwStepLine":7,"gherkinStepLine":11,"keywordType":"Context","textWithKeyword":"Given I visit the Landlord's make a report Page","stepMatchArguments":[]},{"pwStepLine":8,"gherkinStepLine":12,"keywordType":"Action","textWithKeyword":"When I choose to navigate to \"When did the person move into your property?\" page for LMR","stepMatchArguments":[{"group":{"start":24,"value":"\"When did the person move into your property?\"","children":[{"start":25,"value":"When did the person move into your property?","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":9,"gherkinStepLine":15,"keywordType":"Action","textWithKeyword":"And I select continue","stepMatchArguments":[]},{"pwStepLine":10,"gherkinStepLine":16,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":11,"gherkinStepLine":17,"keywordType":"Outcome","textWithKeyword":"And I should see \"Tell us when the person moved into your property\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Tell us when the person moved into your property\"","children":[{"start":14,"value":"Tell us when the person moved into your property","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":12,"gherkinStepLine":19,"keywordType":"Action","textWithKeyword":"When I enter \"Tomorrow's date\" into \"When did the person move into your property?\" field and page for LMR","stepMatchArguments":[{"group":{"start":8,"value":"\"Tomorrow's date\"","children":[{"start":9,"value":"Tomorrow's date","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"},{"group":{"start":31,"value":"\"When did the person move into your property?\"","children":[{"start":32,"value":"When did the person move into your property?","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":13,"gherkinStepLine":20,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":14,"gherkinStepLine":21,"keywordType":"Outcome","textWithKeyword":"And I should see \"This date must not be in the future\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"This date must not be in the future\"","children":[{"start":14,"value":"This date must not be in the future","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":15,"gherkinStepLine":23,"keywordType":"Action","textWithKeyword":"When I enter \"30/11/2014\" into \"When did the person move into your property?\" field and page for LMR","stepMatchArguments":[{"group":{"start":8,"value":"\"30/11/2014\"","children":[{"start":9,"value":"30/11/2014","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"},{"group":{"start":26,"value":"\"When did the person move into your property?\"","children":[{"start":27,"value":"When did the person move into your property?","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":16,"gherkinStepLine":24,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":17,"gherkinStepLine":25,"keywordType":"Outcome","textWithKeyword":"And I should see \"This service only applies to tenancy agreements that started after 30 November 2014\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"This service only applies to tenancy agreements that started after 30 November 2014\"","children":[{"start":14,"value":"This service only applies to tenancy agreements that started after 30 November 2014","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":18,"gherkinStepLine":26,"keywordType":"Action","textWithKeyword":"When I enter \"Today's date\" into \"When did the person move into your property?\" field and page for LMR","stepMatchArguments":[{"group":{"start":8,"value":"\"Today's date\"","children":[{"start":9,"value":"Today's date","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"},{"group":{"start":28,"value":"\"When did the person move into your property?\"","children":[{"start":29,"value":"When did the person move into your property?","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":19,"gherkinStepLine":29,"keywordType":"Action","textWithKeyword":"And I select continue","stepMatchArguments":[]},{"pwStepLine":20,"gherkinStepLine":30,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":21,"gherkinStepLine":31,"keywordType":"Outcome","textWithKeyword":"And I should see \"Enter tenant's full name¬Please enter the tenant's date of birth\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Enter tenant's full name¬Please enter the tenant's date of birth\"","children":[{"start":14,"value":"Enter tenant's full name¬Please enter the tenant's date of birth","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":22,"gherkinStepLine":33,"keywordType":"Action","textWithKeyword":"When I complete Existing tenant's information fields with the below details:","stepMatchArguments":[]},{"pwStepLine":23,"gherkinStepLine":37,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":24,"gherkinStepLine":38,"keywordType":"Outcome","textWithKeyword":"And I should see \"Date of birth entered must reflect 18 years or older\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Date of birth entered must reflect 18 years or older\"","children":[{"start":14,"value":"Date of birth entered must reflect 18 years or older","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":25,"gherkinStepLine":40,"keywordType":"Action","textWithKeyword":"When I complete Existing tenant's information fields with the below details:","stepMatchArguments":[]},{"pwStepLine":26,"gherkinStepLine":44,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":27,"gherkinStepLine":45,"keywordType":"Outcome","textWithKeyword":"And I should see \"Please enter a valid date\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Please enter a valid date\"","children":[{"start":14,"value":"Please enter a valid date","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":28,"gherkinStepLine":46,"keywordType":"Action","textWithKeyword":"When I complete Existing tenant's information fields with the below details:","stepMatchArguments":[]},{"pwStepLine":29,"gherkinStepLine":52,"keywordType":"Action","textWithKeyword":"And I select continue","stepMatchArguments":[]},{"pwStepLine":30,"gherkinStepLine":53,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":31,"gherkinStepLine":54,"keywordType":"Outcome","textWithKeyword":"And I should see \"Enter the property address line 1¬Enter the town/city¬Enter the postcode of the rental property\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Enter the property address line 1¬Enter the town/city¬Enter the postcode of the rental property\"","children":[{"start":14,"value":"Enter the property address line 1¬Enter the town/city¬Enter the postcode of the rental property","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":32,"gherkinStepLine":56,"keywordType":"Action","textWithKeyword":"When I complete Rental property address with the below details:","stepMatchArguments":[]},{"pwStepLine":33,"gherkinStepLine":62,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":34,"gherkinStepLine":63,"keywordType":"Outcome","textWithKeyword":"And I should see \"Your postcode is not in England. If your property is not in England, you do not have to do a right to rent check\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Your postcode is not in England. If your property is not in England, you do not have to do a right to rent check\"","children":[{"start":14,"value":"Your postcode is not in England. If your property is not in England, you do not have to do a right to rent check","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":35,"gherkinStepLine":64,"keywordType":"Action","textWithKeyword":"When I complete Rental property address with the below details:","stepMatchArguments":[]},{"pwStepLine":36,"gherkinStepLine":72,"keywordType":"Action","textWithKeyword":"And I select continue","stepMatchArguments":[]},{"pwStepLine":37,"gherkinStepLine":73,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":38,"gherkinStepLine":74,"keywordType":"Outcome","textWithKeyword":"And I should see \"Enter the landlord's full name¬Enter the landlord's email\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Enter the landlord's full name¬Enter the landlord's email\"","children":[{"start":14,"value":"Enter the landlord's full name¬Enter the landlord's email","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":39,"gherkinStepLine":76,"keywordType":"Action","textWithKeyword":"When I complete Landlord's or agent's information with the below details:","stepMatchArguments":[]},{"pwStepLine":40,"gherkinStepLine":81,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":41,"gherkinStepLine":82,"keywordType":"Outcome","textWithKeyword":"And I should see \"Enter a valid email¬Enter a valid phone number\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Enter a valid email¬Enter a valid phone number\"","children":[{"start":14,"value":"Enter a valid email¬Enter a valid phone number","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":42,"gherkinStepLine":84,"keywordType":"Action","textWithKeyword":"When I complete Landlord's or agent's information with the below details:","stepMatchArguments":[]},{"pwStepLine":43,"gherkinStepLine":89,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":44,"gherkinStepLine":90,"keywordType":"Outcome","textWithKeyword":"And I should see \"Enter a valid email¬Enter a valid phone number\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Enter a valid email¬Enter a valid phone number\"","children":[{"start":14,"value":"Enter a valid email¬Enter a valid phone number","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":45,"gherkinStepLine":92,"keywordType":"Action","textWithKeyword":"When I complete Landlord's or agent's information with the below details:","stepMatchArguments":[]},{"pwStepLine":46,"gherkinStepLine":97,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":47,"gherkinStepLine":98,"keywordType":"Outcome","textWithKeyword":"And I should see \"Enter a valid email\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Enter a valid email\"","children":[{"start":14,"value":"Enter a valid email","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":48,"gherkinStepLine":99,"keywordType":"Action","textWithKeyword":"When I complete Landlord's or agent's information with the below details:","stepMatchArguments":[]},{"pwStepLine":49,"gherkinStepLine":104,"keywordType":"Action","textWithKeyword":"And I select continue","stepMatchArguments":[]},{"pwStepLine":50,"gherkinStepLine":106,"keywordType":"Action","textWithKeyword":"And I select submit report button","stepMatchArguments":[]},{"pwStepLine":51,"gherkinStepLine":107,"keywordType":"Outcome","textWithKeyword":"Then I should see \"There is a problem\" error message displayed","stepMatchArguments":[{"group":{"start":13,"value":"\"There is a problem\"","children":[{"start":14,"value":"There is a problem","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]},{"pwStepLine":52,"gherkinStepLine":108,"keywordType":"Outcome","textWithKeyword":"And I should see \"Confirm you have read the Data Protection statement\" error summary","stepMatchArguments":[{"group":{"start":13,"value":"\"Confirm you have read the Data Protection statement\"","children":[{"start":14,"value":"Confirm you have read the Data Protection statement","children":[{}]},{"children":[{}]}]},"parameterTypeName":"string"}]}]},
-]; // bdd-data-end
