+++ e2e-tests/pages/lmr-make-a-report-page.ts
@@ -1,5 +1,5 @@
 import { Page, Locator } from '@playwright/test';
-import { BasePage } from './BasePage';
+import { BasePage } from './base-page';
 
 export class LmrMakeAReportPage extends BasePage {
