+++ e2e-tests/pages/lmr-privacy-policy-page.ts
@@ -1,5 +1,5 @@
 import { Page, Locator } from '@playwright/test';
-import { BasePage } from './BasePage';
+import { BasePage } from './base-page';
 
 export class LmrPrivacyPolicyPage extends BasePage {
