+++ package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "main": "index.js",
   "engines": {
-    "node": ">=20.18.0 <21.0.0"
+    "node": ">=24.15.0 <25.0.0"
   },
   "scripts": {
     "start": "node server.js",
@@ -14,6 +14,13 @@
     "test:lint": "eslint . --config ./node_modules/eslint-config-hof/default.js",
     "test:unit": "LOG_LEVEL=error nyc _mocha \"test/_unit/**/*.spec.js\"",
     "test:snyk": "snyk config set api=SNYK_TOKEN && snyk test",
+    "bddgen": "bddgen",
+    "test:e2e": "bddgen && playwright test",
+    "test:e2e:smoke": "bddgen && playwright test --grep-invert @history",
+    "test:e2e:history": "bddgen && playwright test --grep @history",
+    "test:e2e:headed": "bddgen && playwright test --headed",
+    "test:e2e:debug": "bddgen && playwright test --debug",
+    "test:e2e:report": "playwright show-report",
     "build": "hof-build",
     "postinstall": "yarn run build"
   },
@@ -32,6 +39,7 @@
     "@playwright/test": "^1.60.0",
     "@types/node": "^25.9.2",
     "chai": "^5.2.1",
+    "dotenv": "^17.4.2",
     "eslint": "8.57.0",
     "eslint-config-hof": "^1.3.4",
     "mocha": "^11.7.6",
