+++ playwright.config.ts
@@ -1,12 +1,19 @@
-import { defineConfig, devices } from '@playwright/test';
+import { defineConfig } from '@playwright/test';
 import { defineBddConfig } from 'playwright-bdd';
+import dotenv from 'dotenv';
+
+dotenv.config({ quiet: true });
+
+const port = Number(process.env.PLAYWRIGHT_PORT || 8080);
+const baseURL = process.env.PLAYWRIGHT_BASE_URL || `http://localhost:${port}`;
 
 const testDir = defineBddConfig({
   features: 'e2e-tests/features/**/*.feature',
   steps: [
     'e2e-tests/steps/**/*.step.ts',
     'e2e-tests/fixture/fixtures.ts'
-  ]
+  ],
+  outputDir: '.features-gen'
 });
 
 export default defineConfig({
@@ -17,7 +24,7 @@ export default defineConfig({
   },
 
   /* Run tests in files in parallel */
-  fullyParallel: false,
+  fullyParallel: true,
   /* Opt out of parallel tests on CI. */
   workers: process.env.CI ? 2 : 1,
 
@@ -28,18 +35,26 @@ export default defineConfig({
   retries: process.env.CI ? 2 : 1,
 
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
-  reporter: process.env.CI ? 'github' : 'html',
+  reporter: [['html', { open: 'never' }], ['list']],
 
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
-    baseURL: 'https://lmr.uat.sas-notprod.homeoffice.gov.uk/',
-    headless: !!process.env.CI,
+    baseURL,
     viewport: null,
     screenshot: 'only-on-failure',
     video: 'retain-on-failure',
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
   },
+  
+  webServer: process.env.PLAYWRIGHT_BASE_URL
+    ? undefined
+    : {
+      command: 'yarn start:dev',
+      port,
+      reuseExistingServer: !process.env.CI,
+      timeout: 120_000
+    },
 
   /* Configure projects for major browsers */
   projects: [
